clean code re-org before re-adding new features
