+++ moesif-gloo-wasm/src/config.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Default, Clone, Serialize, Deserialize)]
+pub struct Config {
+    pub moesif_application_id: String,
+    pub user_id_header: Option<String>,
+    pub company_id_header: Option<String>,
+    #[serde(default)]
+    pub queue_id: u32,
+}
